requirements.txt
demucs
openai-whisper
opencc
colorama